requirements.txt
fastapi==0.99.1
uvicorn==0.22.0
pyyaml==6.0
